início seguir usuário

+++ app/Http/Controllers/FollowedUserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\FollowedUser;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class FollowedUserController extends Controller
 {
@@ -14,12 +15,20 @@ class FollowedUserController extends Controller
 
     public function store(Request $request)
     {
-        FollowedUser::create($request->all());
+        $followedUser = FollowedUser::create($request->all());
+        return response()->json($followedUser, 200);
     }
 
     public function show(string $id)
     {
         return FollowedUser::findOrFail($id);
+    }
+
+    public function showFollowedUser(string $followedUserId)
+    {
+        $followedUser = FollowedUser::where("user_id", Auth::user()->id)->where("followed_user_id", $followedUserId)->first();
+        if ($followedUser) return response()->json($followedUser, 200);
+        return response()->json(false, 200);
     }
 
     public function update(Request $request, string $id)
@@ -28,9 +37,13 @@ class FollowedUserController extends Controller
         $followedUser->update($request->all());
     }
 
-    public function destroy(string $id)
+    public function destroy(string $followedUserId)
     {
-        $followedUser = FollowedUser::findOrFail($id);
-        $followedUser->delete();
+        $followedUser = FollowedUser::where("user_id", Auth::user()->id)->where("followed_user_id", $followedUserId)->first();
+        if ($followedUser) {
+            $followedUser->delete();
+            return response()->json($followedUser, 200);
+        }
+        return response()->json(false, 200);
     }
 }
